+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - uuid option for so ls
 - Duration in so-interval ls command.
 - Period option for so-interval ls command.
-- Interval tab in GUI with equivalent of ls and modify commands.
+- Interval tab in GUI with equivalent of ls, modify, delete and add commands.
 
 ### Removed